refactor(dataframe): take Scalar/enum-class, not raw C types, in the C++ API
The C++ Series/DataFrame/LazyFrame/Expr surface exposed raw C ABI types:
dftu_scalar in clip/fillna/fill_null, and bare int32 op codes in the expr
builders. Switch these to the ergonomic C++ types - Scalar and the existing
PrimOp/UnaryOp/CmpOp/LogicalOp enum classes - so a C++ caller never handles a
C ABI value directly. Add zero-copy span accessors to Series (values(),
offsets_span(), take(span)).

The extern-"C" dftu_expr_* shims keep their int32/dftu_scalar params (the
stable C seam) and cast to the enum/Scalar when calling the C++ builder. Scalar
converts implicitly, so call sites are unchanged.

+++ src/dftracer/utils/trace/views/view_agg_engine.cpp
@@ -894,7 +894,7 @@ coro::CoroTask<EnginePrep> prepare_engine_group(const ViewPlan& plan) {
             ts_d * dataframe::expr_lit(spec.bucket_scale) -
             dataframe::expr_lit(static_cast<double>(spec.bucket_origin));
         dataframe::Expr floored = dataframe::expr_unary(
-            static_cast<std::int32_t>(dataframe::UnaryOp::Floor),
+            dataframe::UnaryOp::Floor,
             rel / dataframe::expr_lit(spec.bucket_interval));
         dataframe::Expr bucket =
             dataframe::expr_cast(dataframe::TypeId::Int64, floored) *
